Ai-Agent-Service/pyproject.toml
[project]
name = "micro-hack-ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.4",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
